backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
neo4j==5.15.0
elasticsearch==8.11.0
requests==2.31.0
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
pydantic==2.5.0
alembic==1.12.1
redis==5.0.1
celery==5.3.4
pytest==7.4.3
httpx==0.25.2

# Multi-modal ingestion
yt-dlp>=2023.3.4
whisper>=1.0.0
scholarly>=1.7.0
arxiv>=1.4.0

# Additional NLP
nltk>=3.8.0

# Premium features
aiohttp>=3.8.0
